Dls release v1.3.0 to DEV

+++ DigitalLearningSolutions.Web/Controllers/FrameworksController/ImportCompetencies.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Linq.Expressions;
 
 namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
 {
@@ -45,7 +46,13 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
         public IActionResult StartImport(ImportCompetenciesFormData model, int frameworkId, string tabname, bool isNotBlank)
         {
             if (!ModelState.IsValid)
-                return View("Developer/Import/Index", model);
+            {
+                var adminId = GetAdminId();
+                var framework = frameworkService.GetFrameworkDetailByFrameworkId(frameworkId, adminId);
+                var viewModel = new ImportCompetenciesViewModel(framework, isNotBlank);
+                viewModel.ImportFile = model.ImportFile;
+                return View("Developer/Import/Index", viewModel);
+            }
             try
             {
                 var adminUserID = User.GetAdminIdKnownNotNull();
@@ -73,7 +80,7 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
             }
             catch (InvalidHeadersException)
             {
-                return View("Developer/Import/ImportFailed");
+                return RedirectToAction("ImportFailed", "Frameworks", new { frameworkId, tabname, isNotBlank });
             }
         }
         [Route("/Framework/{frameworkId}/{tabname}/Import/Uploaded")]
@@ -97,9 +104,18 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
             catch (InvalidHeadersException)
             {
                 FileHelper.DeleteFile(webHostEnvironment, data.CompetenciesFileName);
-                return View("Developer/Import/ImportFailed");
+                return RedirectToAction("ImportFailed", "Frameworks", new { data.FrameworkId, tabname = "Structure", data.IsNotBlank });
             }
         }
+        [Route("/Framework/{frameworkId}/{tabname}/Import/Failed")]
+        public IActionResult ImportFailed(int frameworkId, string tabname, bool isNotBlank)
+        {
+            var adminId = GetAdminId();
+            var framework = frameworkService.GetFrameworkDetailByFrameworkId(frameworkId, adminId);
+            var viewModel = new ImportCompetenciesViewModel(framework, isNotBlank);
+            return View("Developer/Import/ImportFailed", viewModel);
+        }
+
         [Route("/Framework/{frameworkId}/{tabname}/Import/Ordering")]
         public IActionResult ApplyCompetencyOrdering()
         {
@@ -165,15 +181,16 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
         public IActionResult AddAssessmentQuestions(AddAssessmentQuestionsFormData model)
         {
             var data = GetBulkUploadData();
-            data.AddDefaultAssessmentQuestions = model.AddDefaultAssessmentQuestions;
+
             if (model.AddDefaultAssessmentQuestions)
             {
-                data.DefaultQuestionIDs = model.DefaultAssessmentQuestionIDs;
+                data.DefaultQuestionIDs = model.DefaultAssessmentQuestionIDs ?? [];
             }
             else
             {
                 data.DefaultQuestionIDs = [];
             }
+            data.AddDefaultAssessmentQuestions = (data.DefaultQuestionIDs.Count > 0 && model.AddDefaultAssessmentQuestions);
             data.AddCustomAssessmentQuestion = model.AddCustomAssessmentQuestion;
             if (model.AddCustomAssessmentQuestion)
             {
@@ -183,9 +200,8 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
             {
                 data.CustomAssessmentQuestionID = null;
             }
-            if (data.CompetenciesToUpdateCount > 0)
+            if (data.CompetenciesToUpdateCount > 0 && (data.DefaultQuestionIDs.Count + (data.CustomAssessmentQuestionID != null ? 1 : 0) > 0))
             {
-                data.AddAssessmentQuestionsOption = 2;
                 setBulkUploadData(data);
                 return RedirectToAction("AddQuestionsToWhichCompetencies", "Frameworks", new { frameworkId = data.FrameworkId, tabname = data.TabName });
             }
@@ -230,6 +246,7 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
         [Route("/Framework/{frameworkId}/{tabname}/Import/Summary")]
         public IActionResult ImportSummary()
         {
+            if (!TempData.Any()) return  RedirectToAction("StatusCode", "LearningSolutions", new { code = 410 });
             var data = GetBulkUploadData();
             var model = new ImportSummaryViewModel(data);
             return View("Developer/Import/ImportSummary", model);
@@ -245,11 +262,6 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
             var workbook = new XLWorkbook(filePath);
             var results = importCompetenciesFromFileService.ProcessCompetenciesFromFile(workbook, adminId, data.FrameworkId, data.FrameworkVocubulary, data.ReorderCompetenciesOption, data.AddAssessmentQuestionsOption, data.AddCustomAssessmentQuestion ? (int)data.CustomAssessmentQuestionID : 0, data.AddDefaultAssessmentQuestions ? data.DefaultQuestionIDs : []);
             data.ImportCompetenciesResult = results;
-            //TO DO apply ordering changes if required:
-            if (data.ReorderCompetenciesOption == 2 && data.CompetenciesToReorderCount > 0)
-            {
-
-            }
             setBulkUploadData(data);
             return RedirectToAction("UploadResults", "Frameworks", new { frameworkId = data.FrameworkId, tabname = data.TabName });
         }
@@ -263,12 +275,25 @@ namespace DigitalLearningSolutions.Web.Controllers.FrameworksController
             return View("Developer/Import/UploadResults", model);
         }
         [Route("CancelImport")]
-        public IActionResult CancelImport()
+        public IActionResult CancelImport(int? frameworkId)
         {
-            var data = GetBulkUploadData();
-            var frameworkId = data.FrameworkId;
-            FileHelper.DeleteFile(webHostEnvironment, data.CompetenciesFileName);
-            TempData.Clear();
+            try
+            {
+                var data = GetBulkUploadData();
+                frameworkId = data.FrameworkId;
+                if (!string.IsNullOrWhiteSpace(data.CompetenciesFileName))
+                {
+                    FileHelper.DeleteFile(webHostEnvironment, data.CompetenciesFileName);
+                }
+            }
+            catch
+            {
+
+            }
+            finally
+            {
+                TempData.Clear();
+            }
             return RedirectToAction("ViewFramework", new { frameworkId, tabname = "Structure" });
         }
         private void setupBulkUploadData(int frameworkId, int adminUserID, string competenciessFileName, string tabName, bool isNotBlank)
